Move tests to native
