+++ packages/editor/src/store/test/actions.js
@@ -861,14 +861,4 @@ describe( 'Editor actions', () => {
 			} );
 		} );
 	} );
-
-	describe( 'togglePostTitleSelection', () => {
-		it( 'should return the TOGGLE_POST_TITLE_SELECTION action', () => {
-			const result = actions.togglePostTitleSelection( true );
-			expect( result ).toEqual( {
-				type: 'TOGGLE_POST_TITLE_SELECTION',
-				isSelected: true,
-			} );
-		} );
-	} );
 } );
