+++ packages/editor/src/store/test/reducer.js
@@ -18,7 +18,6 @@ import {
 	saving,
 	reusableBlocks,
 	postSavingLock,
-	postTitle,
 	previewLink,
 } from '../reducer';
 import { INITIAL_EDITS_DEFAULTS } from '../defaults';
@@ -845,32 +844,6 @@ describe( 'state', () => {
 			} );
 
 			expect( state ).toEqual( {} );
-		} );
-	} );
-
-	describe( 'postTitle', () => {
-		describe( 'isSelected()', () => {
-			it( 'should not be selected by default', () => {
-				expect( postTitle( undefined, {} ).isSelected ).toBe( false );
-			} );
-
-			it( 'should return false if not selecting the post title', () => {
-				const action = {
-					type: 'TOGGLE_POST_TITLE_SELECTION',
-					isSelected: false,
-				};
-
-				expect( postTitle( { isSelected: true }, action ).isSelected ).toBe( false );
-			} );
-
-			it( 'should return true if selecting the post title', () => {
-				const action = {
-					type: 'TOGGLE_POST_TITLE_SELECTION',
-					isSelected: true,
-				};
-
-				expect( postTitle( { isSelected: false }, action ).isSelected ).toBe( true );
-			} );
 		} );
 	} );
