+++ packages/editor/src/store/test/selectors.js
@@ -66,7 +66,6 @@ const {
 	getPermalink,
 	getPermalinkParts,
 	isPostSavingLocked,
-	isPostTitleSelected,
 	canUserUseUnfilteredHTML,
 } = selectors;
 
@@ -1095,28 +1094,6 @@ describe( 'selectors', () => {
 			};
 
 			expect( isPostSavingLocked( state ) ).toBe( false );
-		} );
-	} );
-
-	describe( 'isPostTitleSelected', () => {
-		it( 'should return true if the post title is selected', () => {
-			const state = {
-				postTitle: {
-					isSelected: true,
-				},
-			};
-
-			expect( isPostTitleSelected( state ) ).toBe( true );
-		} );
-
-		it( 'should return false if the post title is not selected', () => {
-			const state = {
-				postTitle: {
-					isSelected: false,
-				},
-			};
-
-			expect( isPostTitleSelected( state ) ).toBe( false );
 		} );
 	} );
